% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_Net.R
\name{Alpha_Net}
\alias{Alpha_Net}
\title{Alpha_Net}
\usage{
Alpha_Net(Net_Alpha, Net_DevYear, Cumulative.T)
}
\arguments{
\item{Net_Alpha}{Set of alpha factors predicted by each individual RNN.}

\item{Net_DevYear}{Index that allows to identify the development year of each alpha factor.}

\item{Cumulative.T}{Cumulative payments triangle.}
}
\value{
Predictive alpha factors of the MackNet model obtained from the ensemble of RNNs.
}
\description{
It calculates the predictive alpha factors of the RNNs ensemble.
}
