% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA.Build.DB.R
\name{BA.Build.DB}
\alias{BA.Build.DB}
\title{BA.Build.DB}
\usage{
BA.Build.DB(Incremental.T, PI_Ratio)
}
\arguments{
\item{Incremental.T}{Incremental triangle.}

\item{PI_Ratio}{Payments between incurred ratio.}
}
\value{
Database for fitting the MackNet model. It includes the autorregresive component.
}
\description{
It creates the database for fitting the ensemble of RNNs.
}
