% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.Dev.Factor.MACK.R
\name{Boot.Dev.Factor.MACK}
\alias{Boot.Dev.Factor.MACK}
\title{Boot.Ratios.MACK}
\usage{
Boot.Dev.Factor.MACK(Cumulative.T, Link.R)
}
\arguments{
\item{Cumulative.T}{Cumulative triangle.}

\item{Link.R}{Link ratios resampled randomly.}
}
\value{
Development factors resampled randomly.
}
\description{
It generates the set of development factors resampled by using the resampled set of MackNet link ratios.
}
