% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.Residuals.MACK.R
\name{Boot.Residuals.MACK}
\alias{Boot.Residuals.MACK}
\title{Boot.Residuals.MACK}
\usage{
Boot.Residuals.MACK(Incremental.T, Residuals.Boot)
}
\arguments{
\item{Incremental.T}{Original triangle used to fit the MackNet model}

\item{Residuals.Boot}{MackNet residuals.}
}
\value{
Residuals randomly resampled.
}
\description{
It resamples the MackNet residuals.
}
