% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA_TriangleSelectionOrderedRt}
\alias{CA_TriangleSelectionOrderedRt}
\title{Triangle selection made by Meyers (2015).LoB: Comercial Auto}
\description{
Triangle selection made by Meyers (2015).LoB: Comercial Auto
}
\references{
\url{https://www.casact.org/pubs/monographs/papers/01-Meyers.PDF}
}
\keyword{data}
