% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{ChainLaddert.t_1}
\alias{ChainLaddert.t_1}
\title{ChainLaddert.t_1}
\usage{
ChainLaddert.t_1(Triangle.Cumulative, Development.Factors)
}
\arguments{
\item{Triangle.Cumulative}{Cumulative triangle including the payments forecasted in t.}

\item{Development.Factors}{Development factors used within the Chain Ladder technique.}
}
\value{
Full triangle obtained after the Chain Ladder methodology assuming payments in t are true.
}
\description{
It applies the Chain Ladder technique assuming that payments in t are true.
}
