% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_Ultimate_Reference.R
\name{CL_Ultimate_Reference}
\alias{CL_Ultimate_Reference}
\title{CL_Ultimate_Reference}
\usage{
CL_Ultimate_Reference(Cumulative.T, Exposure, PI_Ratio)
}
\arguments{
\item{Cumulative.T}{Cumulative payments triangle.}

\item{Exposure}{Exposure measure.}

\item{PI_Ratio}{Cumulative scaled payments triangle divided between the incurred payments triangle}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{X} Explanatory variables for training the incurred cost MackNet model.
\item \code{Y} Response variable for training the incurred cost MackNet model.
}
}
\description{
It calculates the ultimate from CL to be included within the ensemble of RNNs
}
