% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{CumulativeT.SchedudeP}
\alias{CumulativeT.SchedudeP}
\title{CumulativeT.SchedudeP}
\usage{
CumulativeT.SchedudeP(Row.Data)
}
\arguments{
\item{Row.Data}{Raw data of the selected company.}
}
\value{
Cumulative payments triangle.
}
\description{
It creates the cumulative payments triangle of one company from the Schedule P dataset.
}
