% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{CumulativeT.SchedudeP.Validation}
\alias{CumulativeT.SchedudeP.Validation}
\title{CumulativeT.SchedudeP.Validation}
\usage{
CumulativeT.SchedudeP.Validation(Row.Data)
}
\arguments{
\item{Row.Data}{Raw data of the selected company.}
}
\value{
Cumulative payments upper and lower triangle.
}
\description{
It creates the upper and lower cumulative payments triangle of one company from the Schedule P dataset.
}
