% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB.R
\name{DB}
\alias{DB}
\title{DB}
\usage{
DB(ScaledCumulative.T, ScaledIncurred.T)
}
\arguments{
\item{ScaledCumulative.T}{Cumulative payments triangle divided between the exposure measure.}

\item{ScaledIncurred.T}{Incurred cost triangle divided between the exposure measure.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{train.x} Explanatory variables for training the MackNet model.
\item \code{test.x} Explanatory variables for testing the MackNet model.
\item \code{train.y} Response variable for training the MackNet model.
\item \code{test.y} Response variable for training the MackNet model.
}
}
\description{
It creates the database required to fit the MackNet model.
}
