% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Incurred.R
\name{DB_Incurred}
\alias{DB_Incurred}
\title{DB_Incurred}
\usage{
DB_Incurred(ScaledCumulative.T, ScaledIncurred.T)
}
\arguments{
\item{ScaledCumulative.T}{Cumulative payments triangle divided between the exposure measure.}

\item{ScaledIncurred.T}{Incurred cost triangle divided between the exposure measure.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{train.x} Explanatory variables for training the incurred cost MackNet model.
\item \code{test.x} Explanatory variables for testing the incurred cost MackNet model.
\item \code{train.y} Response variable for training the incurred cost MackNet model.
\item \code{test.y} Response variable for training the incurred cost MackNet model.
}
}
\description{
It creates the database required to fit the incurred cost MackNet model.
}
