% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFactors.t.Predictive.R
\name{DevFactors.t.Predictive}
\alias{DevFactors.t.Predictive}
\title{DevFactors.t.Predictive}
\usage{
DevFactors.t.Predictive(Triangle.Cumulative)
}
\arguments{
\item{Triangle.Cumulative}{Upper and lower triangle used for estimating the predictive development factors.}
}
\value{
Predictive development factors
}
\description{
Calculation of predictive development factors.
}
