% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{DevFactors.t}
\alias{DevFactors.t}
\title{DevFactors.t}
\usage{
DevFactors.t(Triangle.Cumulative)
}
\arguments{
\item{Triangle.Cumulative}{Cumulative upper triangle.}
}
\value{
Development factors.
}
\description{
It calculates the development factors derived from the observed information (Upper triangle).
}
