% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFactors.t1.Predictive.R
\name{DevFactors.t1.Predictive}
\alias{DevFactors.t1.Predictive}
\title{DevFactors.t1.Predictive}
\usage{
DevFactors.t1.Predictive(Triangle.Cumulative)
}
\arguments{
\item{Triangle.Cumulative}{Cumulative triangle}
}
\value{
Predictive development factors assuming the payment forecasted in t is true.
}
\description{
Calculation of the predictive development factors, assuming that the payment forecasted in t is true.
}
