% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFactors_Net.R
\name{DevFactors_Net}
\alias{DevFactors_Net}
\title{DevFactors_Net}
\usage{
DevFactors_Net(RNN_DevFactors, RNN_DevYear, Cumulative.T, OutliersControl)
}
\arguments{
\item{RNN_DevFactors}{Set of development factors predicted by each individual RNN.}

\item{RNN_DevYear}{Index that allows to identify the development year of each development factor.}

\item{Cumulative.T}{Cumulative payments triangle.}

\item{OutliersControl}{Development factors below 0.975 are not allowed when this variable is set to 1.}
}
\value{
Predictive development factors of the MackNet model obtained from the ensemble of RNNs.
}
\description{
It calculates the predictive development factors of the RNNs ensemble.
}
