% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final.Residuals.Mack.Full.R
\name{Final.Residuals.Mack.Full}
\alias{Final.Residuals.Mack.Full}
\title{Final.Residuals.Mack.Full}
\usage{
Final.Residuals.Mack.Full(Residuals, ZeroMean)
}
\arguments{
\item{Residuals}{Scaled bias adjusted residuals.}

\item{ZeroMean}{If this variable is set to 0, residuals are not scaled to have zero mean. By default they are adjusted.}
}
\value{
Scaled bias and mean adjusted MackNet residuals
}
\description{
It computes the scaled bias and mean adjusted MackNet residuals.
}
