% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Full.Cumulative}
\alias{Full.Cumulative}
\title{Full.Cumulative}
\usage{
Full.Cumulative(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Incremental upper and lower triangles.}
}
\value{
Cumulative upper and lower triangles.
}
\description{
It generates a full cumulative triangle taking into consideration the incremental data.
}
