% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full.DevFactors.R
\name{Full.DevFactors}
\alias{Full.DevFactors}
\title{Full.DevFactors}
\usage{
Full.DevFactors(Cumulative.T)
}
\arguments{
\item{Cumulative.T}{Upper and lower triangle used for estimating the predictive development factors.}
}
\value{
Full development factors
}
\description{
Calculation of development factors taking into consideration both, all the observed and predicted data.
}
