% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Full.Incremental}
\alias{Full.Incremental}
\title{Full.Incremental}
\usage{
Full.Incremental(Triangle.Cumulative)
}
\arguments{
\item{Triangle.Cumulative}{Cumulative upper and lower triangles.}
}
\value{
Incremental upper and lower triangles.
}
\description{
It generates a full incremental triangle taking into consideration the cumulative data.
}
