% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{IncurredT.SchedudeP}
\alias{IncurredT.SchedudeP}
\title{IncurredT.SchedudeP}
\usage{
IncurredT.SchedudeP(Row.Data)
}
\arguments{
\item{Row.Data}{Raw data of the selected company.}
}
\value{
Incurred cost triangle.
}
\description{
It creates the incurred cost triangle of one company from the Schedule P dataset.
}
