% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{IncurredT.SchedudeP.Validation}
\alias{IncurredT.SchedudeP.Validation}
\title{IncurredT.SchedudeP.Validation}
\usage{
IncurredT.SchedudeP.Validation(Row.Data)
}
\arguments{
\item{Row.Data}{Raw data of the selected company.}
}
\value{
Incurred cost upper and lower triangle.
}
\description{
It creates the upper and lower incurred cost triangle of one company from the Schedule P dataset.
}
