% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual.backtest.R
\name{Individual.backtest}
\alias{Individual.backtest}
\title{Individual.backtest}
\usage{
Individual.backtest(TrianglesBackup, Validation.T)
}
\arguments{
\item{TrianglesBackup}{Triangles sampled by the model.}

\item{Validation.T}{Upper and lower triangle. Needed to validate the model.}
}
\value{
Percentile of the observed data vs the distribution generated for every AY/DY combination
}
\description{
It computes the percentile of the observed data vs the distribution generated for every AY/DY combination
}
