% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{ins.line.data}
\alias{ins.line.data}
\title{ins.line.data}
\usage{
ins.line.data(g.code, a)
}
\arguments{
\item{g.code}{Company code}

\item{a}{Schedule P Database}
}
\value{
Raw data from the selected company code
}
\description{
It obtains the raw data from Schedule P Data
}
