% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastDiagonal.R
\name{LastDiagonal}
\alias{LastDiagonal}
\title{LastDiagonal}
\usage{
LastDiagonal(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Triangle}
}
\value{
Aggregated value of the last diagonal of the triangle.
}
\description{
It obtains the value of the last diagonal of the triangle.
}
