% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastDiagonal.Vector.R
\name{LastDiagonal.Vector}
\alias{LastDiagonal.Vector}
\title{LastDiagonal.Vector}
\usage{
LastDiagonal.Vector(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Triangle}
}
\value{
Vector with all the values of the last diagonal of the triangle.
}
\description{
It obtains a vector with all the values of the last diagonal of the triangle.
}
