% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoB.Backtest.R
\name{LoB.Backtest}
\alias{LoB.Backtest}
\title{LoB.Backtest}
\usage{
LoB.Backtest(Reservest.D, Reservest.D.Vector, Reserves.Validation)
}
\arguments{
\item{Reservest.D}{Total reserves sampled by the model.}

\item{Reservest.D.Vector}{Total reserves by AY sampled by the model.}

\item{Reserves.Validation}{Actual reserve.}
}
\value{
Percentile of the observed data vs the distribution generated by AY and totals.
}
\description{
It computes the percentile of the observed data vs the distribution generated by AY and totals.
}
