% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mack.Bias.Adj.Full.R
\name{Mack.Bias.Adj.Full}
\alias{Mack.Bias.Adj.Full}
\title{Mack.Bias.Adj.Full}
\usage{
Mack.Bias.Adj.Full(Pearson.Resid, Constant.Param)
}
\arguments{
\item{Pearson.Resid}{MackNet residuals}

\item{Constant.Param}{If this variable is set to 0, the bias adjustment suggested by England y Verrall (2006) in "Predictive Distributions of Outstanding Liabilities in General Insurance" is applied, this means that residuals are multiplied by N/(N-p). In case this variable is set to 1, the adjustment suggested by Mack (1993) in "Distribution-free calculation of the standard error of chain ladder reserve estimates" is applied, this means that residuals are multiplied by n(i)/(n(i)-1). Finally, if this variable is set to 2, no adjustment is applied.}
}
\value{
Bias adjustment to be applied to the MackNet residuals.
}
\description{
It calculates the bias adjustment to be applied to the residuals.
}
