% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackNet_Boot.R
\name{MackNet_Boot}
\alias{MackNet_Boot}
\title{MackNet_Boot}
\usage{
MackNet_Boot(Simulations, Parameters, Cumulative.T, OutliersControl)
}
\arguments{
\item{Simulations}{Number of samples produced by the MackNet model}

\item{Parameters}{List with all the MackNet parameters. Please refer to function MackNet_Parameters for further details}

\item{Cumulative.T}{Cumulative payments triangle.}

\item{OutliersControl}{Development factors below 0.975 are not allowed when this variable is set to 1.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{Residuals_Boot} Residuals used within the bootstrap procedure.
\item \code{Payments.D} It contains the total following year payments sampled by the MackNet model.
\item \code{Reservest.D} It contains the total reserves sampled by the MackNet model.
\item \code{Reservest_1.D} It contains the total reserves in t+1 sampled by the MackNet model.
\item \code{Ultimatest.D} It contains the ultimate cost sampled by the MackNet model.
\item \code{Payments.D.Vector} It contains the following year payments by accident year sampled by the MackNet model.
\item \code{Reservest.D.Vector} It contains the reserves by accident year sampled by the MackNet model.
\item \code{Reservest_1.D.Vector} It contains the reserves in t+1 by accident year sampled by the MackNet model.
\item \code{Ultimatest.D.Vector} It contains the ultimate cost by accident year sampled by the MackNet model.
\item \code{TrianglesBackup} This array contains all the triangles sampled by the MackNet model.
\item \code{UpperTrianglesBackup} Only for testing purposes. It assumes that the only observed value is the one included in the first development year.
\item \code{Boot_DevFactors} Predictive development factors sampled by the boostrap procedure.
\item \code{Development_Factors_MackNet} Predictive development factors of the MackNet model.
\item \code{Alpha} Predictive alpha by development year of the MackNet model. This variable is named as sigma in "Mack-Net model: Blending Mack's model with Recurrent Neural Networks".
}
}
\description{
It produces the boostrapping to obtain a full reserves distribution with the MackNet model.
}
