% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackNet_Fit_Incurred.R
\name{MackNet_Fit_Incurred}
\alias{MackNet_Fit_Incurred}
\title{MackNet_Fit_Incurred}
\usage{
MackNet_Fit_Incurred(
  Cumulative.T,
  Incurred.T,
  Exposure,
  AR,
  Ensemble,
  wd,
  Learning,
  drop,
  Epochs,
  MinimumEpochs,
  ES,
  Output
)
}
\arguments{
\item{Cumulative.T}{Cumulative payments triangle.}

\item{Incurred.T}{Incurred cost triangle.}

\item{Exposure}{Exposure measure. Written premiums is an appropriate measure to scale cumulative payments and incurred cost.}

\item{AR}{This variable allows to remove the autorregressive component from the MackNet model when it is set to 0.}

\item{Ensemble}{Number of RNNs included in the ensemble.}

\item{wd}{The optimization algorithm used is ADAM. This variable defines the weighted decay value.}

\item{Learning}{Learning rate.}

\item{drop}{Dropout regularization.}

\item{Epochs}{Maximum number of epochs.}

\item{MinimumEpochs}{Minimum number of epochs.}

\item{ES}{Early Stopping object defined under the keras framework.}

\item{Output}{Linear or ReLU activation function for the output layer.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{TrianglesBackup} Full triangles predicted by each RNN included within the ensemble.
\item \code{Error} Test error.
}
}
\description{
This function fits the ensemble of RNNs required for the incurred cost MackNet model. The optimum weigthed decay is obtained by selecting the configuration that minimizes the test error.
}
