% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MackNet_Parameters.R
\name{MackNet_Parameters}
\alias{MackNet_Parameters}
\title{MackNet_Parameters}
\usage{
MackNet_Parameters(Triangles, Error, Exposure, MackBias, ZeroMean)
}
\arguments{
\item{Triangles}{Full triangles predicted by each RNN included within the ensemble.}

\item{Error}{Test error of all the possible levels of weighted decay.}

\item{Exposure}{Exposure measure. Written premiums is an appropriate measure to scale cumulative payments and incurred cost.}

\item{MackBias}{If this variable is set to 0, the bias adjustment suggested by England y Verrall (2006) in "Predictive Distributions of Outstanding Liabilities in General Insurance" is applied, this means that residuals are multiplied by N/(N-p). In case this variable is set to 1, the adjustment suggested by Mack (1993) in "Distribution-free calculation of the standard error of chain ladder reserve estimates" is applied, this means that residuals are multiplied by n(i)/(n(i)-1). Finally, if this variable is set to 2, no adjustment is applied.}

\item{ZeroMean}{If this variable is set to 0, residuals are not scaled to have zero mean. By default they are adjusted.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{Factors} Predictive development factors of the MackNet model.
\item \code{Alpha} Predictive alpha by development year of the MackNet model. This variable is named as sigma in "Mack-Net model: Blending Mack's model with Recurrent Neural Networks".
\item \code{RNN_Residuals} Residuals of the MackNet model. These residuals are used within the boostrap procedure.
\item \code{Error} Optimum level of weighted decay.
}
}
\description{
It calculated the predictive MackNet parameters.
}
