% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainResults_Incurred.R
\name{MainResults_Incurred}
\alias{MainResults_Incurred}
\title{MainResults}
\usage{
MainResults_Incurred(
  CompanyCode,
  CumulativeValidation,
  IncurredValidation,
  Model
)
}
\arguments{
\item{CompanyCode}{Company code as defined by the Schedule P database}

\item{CumulativeValidation}{Validation triangle.}

\item{IncurredValidation}{Incurred Validation triangle.}

\item{Model}{Paid MackNet model.}
}
\value{
It returns a matrix with following information: \itemize{
\item \code{Company Code} Company code of Schedule P Database.
\item \code{Obs. Reserve} Reserve observed within the Schedule P Database.
\item \code{Obs. Ultimate} Ultimate observed within the Schedule P Database.
\item \code{Est. Ultimate} Ultimate estimated by the paid MackNet model.
\item \code{SE Ultimate} Squared error of the ultimates estimated by the MackNet model.
\item \code{Percentile} It compares the distribution sampled by the paid MackNet model with the reserve actually observed. The percentile where this last value is located within the stochastic process is given.
}
}
\description{
This function collects the ultimate and reserve from the Schedule P Database paid triangles. Additionally, paid MackNet results are reported.
}
