% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_II_SchedulePDatabase.R
\name{NetEarned.SchedudeP}
\alias{NetEarned.SchedudeP}
\title{NetEarned.SchedudeP}
\usage{
NetEarned.SchedudeP(Row.Data)
}
\arguments{
\item{Row.Data}{Raw data of the selected company.}
}
\value{
Premiums by accident year.
}
\description{
It generates a vector with the premiums by accident year of one company from the Schedule P dataset.
}
