% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Payment}
\alias{Payment}
\title{Payment}
\usage{
Payment(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Upper and lower incremental triangle.}
}
\value{
Payments in t.
}
\description{
It calculates the payments in t.
}
