% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Payment.Vector}
\alias{Payment.Vector}
\title{Payment.Vector}
\usage{
Payment.Vector(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Upper and lower incremental triangle.}
}
\value{
Payments by accident year in t.
}
\description{
It calculates the payments by accident year in t.
}
