% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson.Bias.MACK.Full.R
\name{Pearson.Bias.MACK.Full}
\alias{Pearson.Bias.MACK.Full}
\title{Pearson.Bias.MACK.Full}
\usage{
Pearson.Bias.MACK.Full(Pearson.Biasadj, Scale.Param)
}
\arguments{
\item{Pearson.Biasadj}{Bias adjusted residuals.}

\item{Scale.Param}{MackNet scale parameter}
}
\value{
Scaled bias adjusted MackNet residuals
}
\description{
It computes the scaled bias adjusted MackNet residuals.
}
