% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pearson.Residuals.Mack.Full.R
\name{Pearson.Residuals.Mack.Full}
\alias{Pearson.Residuals.Mack.Full}
\title{Pearson.Residuals.Mack.Full}
\usage{
Pearson.Residuals.Mack.Full(Cumulative.T, Dev.Factors)
}
\arguments{
\item{Cumulative.T}{Data used for calculating the residuals.}

\item{Dev.Factors}{Development factors used for calculating the residuals}
}
\value{
Residuals of the MackNet model. These residuals are used within the boostrap procedure.
}
\description{
It generates the residuals needed for the boostrap implementation of MackNet
}
