% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Reserve.t}
\alias{Reserve.t}
\title{Reserve.t}
\usage{
Reserve.t(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Upper and lower incremental triangle.}
}
\value{
Reserve evaluated in t.
}
\description{
It calculates the reserve.
}
