% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Reserve.t_1.Vector}
\alias{Reserve.t_1.Vector}
\title{Reserve.t_1.Vector}
\usage{
Reserve.t_1.Vector(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Upper and lower incremental triangle.}
}
\value{
Reserve by accident year evaluated in t+1.
}
\description{
It calculates the reserve in t+1 by accident year.
}
