% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals_Full.R
\name{Residuals_Full}
\alias{Residuals_Full}
\title{Residuals_Full}
\usage{
Residuals_Full(Triangles, ExposureVector, MackBias, ZeroMean)
}
\arguments{
\item{Triangles}{Data used for calculating the residuals.}

\item{ExposureVector}{Exposure measure. Written premiums is an appropriate measure to scale cumulative payments and incurred cost.}

\item{MackBias}{If this variable is set to 0, the bias adjustment suggested by England y Verrall (2006) in "Predictive Distributions of Outstanding Liabilities in General Insurance" is applied, this means that residuals are multiplied by N/(N-p). In case this variable is set to 1, the adjustment suggested by Mack (1993) in "Distribution-free calculation of the standard error of chain ladder reserve estimates" is applied, this means that residuals are multiplied by n(i)/(n(i)-1). Finally, if this variable is set to 2, no adjustment is applied.}

\item{ZeroMean}{If this variable is set to 0, residuals are not scaled to have zero mean. By default they are adjusted.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{DF} Predictive development factors of the MackNet model.
\item \code{Alpha} Predictive alpha by development year of the MackNet model.
\item \code{Residuals} Residuals of the MackNet model. These residuals are used within the boostrap procedure.
}
}
\description{
It generates the scaled bias-adjusted residuals and the parameters needed for the boostrap implementation of MackNet
}
