% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNN.Triangle.NoAR.R
\name{RNN.Triangle.NoAR}
\alias{RNN.Triangle.NoAR}
\title{RNN.Triangle.NoAR}
\usage{
RNN.Triangle.NoAR(CumulativeScaled.T, PI_Ratio, model)
}
\arguments{
\item{CumulativeScaled.T}{Cumulative payments triangle divided between the exposure measure.}

\item{PI_Ratio}{Observed cumulative payments between incurred cost by development year.}

\item{model}{Keras object containing the structure and optimization algorithm of the RNN.}
}
\value{
Triangle predicted by an individual RNN. The autorregressive component of the MackNet model is not taken into consideration.
}
\description{
It generates the triangle predicted by an individual RNN. The autorregressive component of the MackNet model is not taken into consideration.
}
