% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNN.Triangle.R
\name{RNN.Triangle}
\alias{RNN.Triangle}
\title{RNN.Triangle}
\usage{
RNN.Triangle(CumulativeScaled.T, PI_Ratio, model)
}
\arguments{
\item{CumulativeScaled.T}{Cumulative payments triangle divided between the exposure measure}

\item{PI_Ratio}{Observed cumulative payments between incurred cost by development year.}

\item{model}{Keras object containing the structure and optimization algorithm of the RNN.}
}
\value{
Triangle predicted by an individual RNN.
}
\description{
It generates the triangle predicted by an individual RNN.
}
