% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNN_Keras.R
\name{RNN_Keras}
\alias{RNN_Keras}
\title{RNN_Keras}
\usage{
RNN_Keras(Shape1, Shape2, Learning, wd, drop)
}
\arguments{
\item{Shape1}{Length of the input matrix containing the explanatory variables}

\item{Shape2}{Number of previous observation of the input matrix containing the explanatory variables}

\item{Learning}{Learning rate of the RNN.}

\item{wd}{Weighted decay used within the ADAM optimization algorithm.}

\item{drop}{Level of dropout regularization.}
}
\value{
Keras object containing the structure and optimization algorithm of the RNN.
}
\description{
It creates the keras object to fit each individual RNN.
}
