% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{rotate_clockwise}
\alias{rotate_clockwise}
\title{rotate_clockwise}
\usage{
rotate_clockwise(x)
}
\arguments{
\item{x}{Matrix to be rotated}
}
\value{
Rotated matrix (Clockwise).
}
\description{
It rotates a matrix (Clockwise).
}
