% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{rotate_counter_clockwise}
\alias{rotate_counter_clockwise}
\title{rotate_counter_clockwise}
\usage{
rotate_counter_clockwise(x)
}
\arguments{
\item{x}{Matrix to be rotated}
}
\value{
Rotated matrix (Counter clockwise).
}
\description{
It rotates a matrix (Counter clockwise).
}
