% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale.Parameter.Mack.Full.R
\name{Scale.Parameter.Mack.Full}
\alias{Scale.Parameter.Mack.Full}
\title{Scale.Parameter.Mack.Full}
\usage{
Scale.Parameter.Mack.Full(Pearson.Biasadj)
}
\arguments{
\item{Pearson.Biasadj}{Bias adjusted residuals.}
}
\value{
MackNet scale parameter
}
\description{
It computes the parameter to scale the residuals.
}
