% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Set.LowTriangle.0.KeepDiag}
\alias{Set.LowTriangle.0.KeepDiag}
\title{Set.LowTriangle.0.KeepDiag}
\usage{
Set.LowTriangle.0.KeepDiag(Triangle)
}
\arguments{
\item{Triangle}{Matrix to be modified.}
}
\value{
Original matrix with lower triangle set to 0. Diagonal is kept.
}
\description{
Lower triangle of a matrix is set to 0. Diagonal is kept.
}
