% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{SetDiag.0}
\alias{SetDiag.0}
\title{SetDiag.0}
\usage{
SetDiag.0(Triangle)
}
\arguments{
\item{Triangle}{Matrix to be modified.}
}
\value{
Original matrix with the diagonal set to 0.
}
\description{
Diagonal of a matrix is set to 0.
}
