% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_III_SummaryResults.R
\name{Summary_Results}
\alias{Summary_Results}
\title{Summary_Results}
\usage{
Summary_Results(
  Cumulative.T,
  Original.Cumulative.T.Val,
  Reservest.D,
  Reservest_1.D,
  Payments.D,
  Reservest.D.Vector,
  Payments.D.Vector,
  Reservest_1.D.Vector,
  Ultimatest.D.Vector,
  Ultimatest.D,
  TrianglesBackup,
  Simulations
)
}
\arguments{
\item{Cumulative.T}{Cumulative payments triangle.}

\item{Original.Cumulative.T.Val}{Upper and lower cumulative payment triangle. This information is needed to obtain the error generated by the paid MackNet model.}

\item{Reservest.D}{It contains the total reserves sampled by the MackNet model.}

\item{Reservest_1.D}{It contains the total reserves in t+1 sampled by the MackNet model.}

\item{Payments.D}{It contains the total following year payments sampled by the MackNet model.}

\item{Reservest.D.Vector}{It contains the reserves by accident year sampled by the MackNet model.}

\item{Payments.D.Vector}{It contains the following year payments by accident year sampled by the MackNet model.}

\item{Reservest_1.D.Vector}{It contains the reserves by accident year sampled by the MackNet model.}

\item{Ultimatest.D.Vector}{It contains the ultimate cost by accident year sampled by the MackNet model.}

\item{Ultimatest.D}{It contains the ultimate cost sampled by the MackNet model.}

\item{TrianglesBackup}{This array contains all the triangles sampled by the MackNet model.}

\item{Simulations}{Number of triangle samples to be produced by the MackNet model.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{CDR} Claims Development Result distribution function.
\item \code{CDR_Convergence} Mean convergence of the Claim Development Result
\item \code{Summary} Summary of the results obtained
}
}
\description{
It generates a summary of the results obtained by the paid MackNet model.
}
