% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_III_SummaryResults.R
\name{Summary_Results_Incurred}
\alias{Summary_Results_Incurred}
\title{Summary_Results_Incurred}
\usage{
Summary_Results_Incurred(
  Cumulative.T,
  Original.Cumulative.T.Val.Incurred,
  Reservest.D,
  Reservest.D.Vector,
  Ultimatest.D,
  Ultimatest.D.Vector,
  Ultimatest_1.D,
  Ultimatest_1.D.Vector,
  TrianglesBackup,
  Simulations
)
}
\arguments{
\item{Cumulative.T}{Cumulative payments triangle.}

\item{Original.Cumulative.T.Val.Incurred}{Upper and lower incurred cost triangle. This information is needed to obtain the error generated by the paid MackNet model.}

\item{Reservest.D}{It contains the total reserves sampled by the MackNet model.}

\item{Reservest.D.Vector}{It contains the reserves by accident year sampled by the MackNet model.}

\item{Ultimatest.D}{It contains the ultimate cost sampled by the MackNet model.}

\item{Ultimatest.D.Vector}{It contains the ultimate cost by accident year sampled by the MackNet model.}

\item{Ultimatest_1.D}{It contains the ultimate cost in t+1 sampled by the MackNet model.}

\item{Ultimatest_1.D.Vector}{It contains the ultimate cost in t+1 by accident year sampled by the MackNet model.}

\item{TrianglesBackup}{This array contains all the triangles sampled by the MackNet model.}

\item{Simulations}{Number of triangle samples to be produced by the MackNet model.}
}
\value{
The formula generates the following outputs: \itemize{
\item \code{CDR} Claims Development Result distribution function.
\item \code{CDR_Convergence} Mean convergence of the Claim Development Result
\item \code{Summary} Summary of the results obtained
}
}
\description{
It generates a summary of the results obtained by the incurred MackNet model.
}
