% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.index.R
\name{Test.index}
\alias{Test.index}
\title{Test.index}
\usage{
Test.index(Triangle)
}
\arguments{
\item{Triangle}{incurred cost or cumulative payments triangle.}
}
\value{
Index to identify the observations located in the last diagonal.
}
\description{
It creates the index to identify the last diagonal, which is used as test set.
}
