% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Triangle.Cumulative}
\alias{Triangle.Cumulative}
\title{Triangle.Cumulative.}
\usage{
Triangle.Cumulative(Triangle.Incremental)
}
\arguments{
\item{Triangle.Incremental}{Incremental triangle.}
}
\value{
Cumulative triangle.
}
\description{
Cumulative from incremental triangle.
}
