% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFunctions_I_Results_ChainLadder.R
\name{Triangle.Incremental}
\alias{Triangle.Incremental}
\title{Triangle.Incremental.}
\usage{
Triangle.Incremental(Triangle.Cumulative)
}
\arguments{
\item{Triangle.Cumulative}{Cumulative triangle}
}
\value{
Incremental triangle.
}
\description{
Incremental from cumulative triangle.
}
