% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upper.Triangle.Mack.R
\name{Upper.Triangle.Mack}
\alias{Upper.Triangle.Mack}
\title{Upper.Triangle.Mack}
\usage{
Upper.Triangle.Mack(Cumulative.T, Boot.Ratios)
}
\arguments{
\item{Cumulative.T}{Cumulative triangle.}

\item{Boot.Ratios}{Development factors resampled randomly.}
}
\value{
Resampled triangle assumin that the only observed values are the ones included in the first development year.
}
\description{
Only for testing purposes. It assumes that the only observed values are the ones included in the first development year.
}
