% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wkcomp_pos}
\alias{wkcomp_pos}
\title{Schedule P database of the NAIC Annual Statement was collected from property and casualty insurers that underwrite business in the US and it contains both paid and incurred losses (net of reinsurance) of the accident years going from 1988 to 1997. LoB: Workers Compensation.}
\description{
Schedule P database of the NAIC Annual Statement was collected from property and casualty insurers that underwrite business in the US and it contains both paid and incurred losses (net of reinsurance) of the accident years going from 1988 to 1997. LoB: Workers Compensation.
}
\references{
\url{https://www.casact.org/research/index.cfm?fa=loss_reserves_data}
}
\keyword{data}
